La til dotnet-reportgenerator-globaltool i dotnet-tools.json og la til noen enkle sjekker på om mapper finnes før de slettes i CodeCoverage.ps1 så man slipper diverse feilmeldinger i konsollet.

+++ 2022-04-01-Stryker/Kalkulator.Tests/CodeCoverage.ps1
@@ -2,13 +2,15 @@
 # Source: https://stackoverflow.com/a/70321555/495455
 
 # This only needs to be installed once (globally), if installed it fails silently: 
-dotnet tool install -g dotnet-reportgenerator-globaltool
+#dotnet tool install -g dotnet-reportgenerator-globaltool
 
 # Save currect directory into a variable
 $dir = pwd
 
-# Delete previous test run results (there's a bunch of subfolders named with guids)
-Remove-Item -Recurse -Force $dir/TestResults/
+if (Test-Path $dir/TestResults/) {
+	# Delete previous test run results (there's a bunch of subfolders named with guids)
+	Remove-Item -Recurse -Force $dir/TestResults/
+}
 
 # Run the Coverlet.Collector (this is an NuGet included with XUnit Test Projects)
 $output = [string] (& dotnet test --collect:"XPlat Code Coverage" 2>&1)
@@ -23,8 +25,10 @@ $j = $output.LastIndexOf("coverage")
 $cmdGuid = $output.SubString($i,$j - $i - 1)
 Write-Host $cmdGuid 
 
-# Delete previous test run reports - note if you're getting wrong results do a Solution Clean and Rebuild to remove stale DLLs in the bin folder
-Remove-Item -Recurse -Force $dir/coveragereport/
+if (Test-Path $dir/coveragereport/) {
+	# Delete previous test run reports - note if you're getting wrong results do a Solution Clean and Rebuild to remove stale DLLs in the bin folder
+	Remove-Item -Recurse -Force $dir/coveragereport/
+}
 
 # To keep a history of the Code Coverage we need to use the argument:
 # -historydir:SOME_DIRECTORY 
@@ -33,7 +37,7 @@ if (!(Test-Path -path $dir/CoverageHistory)) {
 }
 
 # Generate the Code Coverage HTML Report
-reportgenerator -reports:"$dir/TestResults/$cmdGuid/coverage.cobertura.xml" -targetdir:"coveragereport" -reporttypes:Html -historydir:$dir/CoverageHistory 
+dotnet reportgenerator -reports:"$dir/TestResults/$cmdGuid/coverage.cobertura.xml" -targetdir:"coveragereport" -reporttypes:Html -historydir:$dir/CoverageHistory 
 
 # Open the Code Coverage HTML Report (if running on a WorkStation)
 $osInfo = Get-CimInstance -ClassName Win32_OperatingSystem
